api update

+++ web/application/models/adminModel.php
@@ -195,19 +195,26 @@ class adminModel extends database
     public function deleteNotice($id)
     {
         $link="";
-        if($this->Query("SELECT url from post_attachements where post_id=",[$id]))
+        if($this->Query("SELECT url from post_attachements where post_id=?",[$id]))
         {
-            $obj=$this->fetch();
-            $link=$obj->url;
-
-        }       
-        // since post has an attachement first delete the attachement then the post, because if we delete the post first there will be no match in post_attachement table for given post
-       if($this->Query("DELETE from post_attachments where post_id=?",[$id])){
+            if($this->rowCount()>0)
+            {
+                $obj=$this->fetch();
+                $link=$obj->url;
+                if($this->Query("DELETE from post_attachments where post_id=?",[$id]))
+                { 
+                    unlink('../../public/assets/dbimages/'.$link);
+                }
+            }     
+        }   
+        if($this->Query("DELETE from post where id=?",[$id]))
+        { 
+            if($this->rowCount()>0){
+                return true; 
+            }  
+        }   
         
-        if($this->Query("DELETE from post where id=?",[$id])){
-            unlink('../../public/assets/dbimages/'.$link);
-        }
-       }
+        return false; 
     }
 
 }
